HANDWRITING DIGIT RECOGNIZATION/requirements.txt
streamlit
tensorflow
numpy
pillow

